--- webapp/src/components/login/dashboard.tsx
@@ -1,16 +1,0 @@
-
-
-import {useAuth} from "../../context/AuthContext";
-
-export function Dashboard(){
-
-    const {getCurrentUser} = useAuth();
-
-    return (<div>
-
-        <h1>DASHBOARD</h1>
-        <p>You are logged in as: </p>
-        <p>{getCurrentUser()?.email}</p>
-
-    </div>);
-}
