--- webapp/src/components/login/protectedRoute.tsx
@@ -1,5 +1,0 @@
-import {Routes,Route} from 'react-router-dom';
-
-export function ProtectedRoute(){
-    return <div>prot</div>
-}
