+++ webapp/src/components/pages/DashboardPage/dashboard.scss
@@ -1,0 +1,7 @@
+.login-page-container {
+    height: calc(100vh - var(--topMenuHeight));
+    background-color: var(--dark1);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
