+++ webapp/src/components/pages/LoginPage/LoginPage.scss
@@ -36,6 +36,7 @@
     box-shadow: 0px 0px 1px 1px rgba(0, 0, 0, 0.103);
     border-radius: 0.5rem;
     transition: 0.4s;
+    color: black;
   }
   input:focus {
     border: none;
