+++ webapp/src/components/pages/RegisterPage/RegisterPage.scss
@@ -1,4 +1,89 @@
 .register-page-container {
   height: calc(100vh - var(--topMenuHeight));
   background-color: var(--dark1);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.register-wrapper {
+  box-shadow: 0px 0px 3rem rgba(255, 255, 255, 0.192);
+  border-radius: 1rem;
+  display: flex;
+  flex-direction: column;
+  width: 30rem;
+  height: 28rem;
+  background-color: var(--white1);
+  h2 {
+    color: var(--orange2);
+    filter: drop-shadow(0px 0px 3px);
+    text-align: center;
+    letter-spacing: var(--letterSpacing);
+    font-size: 2.5rem;
+  }
+  h3,label {
+    color: black;
+    margin: 0;
+    margin-left: 4.5rem;
+  }
+  input {
+    align-self: center;
+    width: 20rem;
+    font-size: 1.2rem;
+    background-color: var(--white);
+    padding: 0.5rem;
+    margin: 0.5rem;
+    box-shadow: 0px 0px 1px 1px rgba(0, 0, 0, 0.103);
+    border-radius: 0.5rem;
+    transition: 0.4s;
+    color: black;
+  }
+  input:focus {
+    border: none;
+    box-shadow: 0px 0px 8px 1px rgba(0, 0, 0, 0.103);
+  }
+}
+
+.buttons {
+  margin-top: 1rem;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+
+  button {
+    cursor: pointer;
+    font-family: "Montserrat-Bold";
+    background-color: var(--orange2);
+    font-size: 1.3rem;
+    padding: 1rem;
+    padding-left: 1.7rem;
+    padding-right: 1.7rem;
+
+    border-radius: 0.5rem;
+    transition: 0.35s;
+  }
+  button:hover {
+    filter: drop-shadow(0px 0px 7px var(--orange2));
+  }
+}
+
+h3:nth-child(4) {
+  margin-top: 1.5rem;
+}
+
+.register-text {
+  font-family: "Montserrat-Regular";
+  margin-top: 1.5rem;
+  text-align: center;
+  color: black;
+  b {
+    cursor: pointer;
+    text-decoration: underline 1px solid black;
+    transition: 0.2s ease-in-out;
+  }
+  b:hover {
+    color: var(--orange2);
+    text-decoration: underline 1px solid var(--orange2);
+  }
 }
